requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
PeakUtils==1.3.5
pillow==11.3.0
pyFirmata==1.1.0
pylsl==1.17.6
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
pysiology==0.0.9.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.3
six==1.17.0
pandas==2.3.3